requirements_dev.txt
fastapi==0.103.1
python-dotenv==1.0.0
uvicorn==0.23.2
pydantic-settings==2.0.3
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0
SQLAlchemy==2.0.21
psycopg2==2.9.7
alembic==1.12.0
redis==5.0.0
fastapi-cache2==0.2.1
asyncpg==0.28.0
passlib[bcrypt]